adding VisionReader for common vision components
