+++ allennlp/data/dataset_readers/__init__.py
@@ -20,6 +20,7 @@ from allennlp.data.dataset_readers.sharded_dataset_reader import ShardedDatasetR
 from allennlp.data.dataset_readers.text_classification_json import TextClassificationJsonReader
 
 try:
+    from allennlp.data.dataset_readers.vision_reader import VisionReader
     from allennlp.data.dataset_readers.vqav2 import VQAv2Reader
     from allennlp.data.dataset_readers.visual_entailment import VisualEntailmentReader
 except ModuleNotFoundError as err:
